requirements.txt
python-dotenv~=1.0.0
requests==2.31.0
img2pdf==0.4.4
fpdf==1.7.2
pypdf==3.12.0
yadisk~=1.3.3